move AnchoredClock to top level
